Separate parsing helpers from general utilities
